Explicitly specify AWS region for Lambda deploys (#129)
This has been pulled from the AWS_REGION / AWS_DEFAULT_REGION environment variables, however this approach is more flexible. It makes it possible to inject the region through code operating at various levels of abstraction, rather than programmatically setting an environment variable, which is messy.

An alternative could be to make the region optional, which would magically work some of the time, however making it explicit is simpler and it feels more Pythonic to provide one obvious way to do it. (I suppose passing `None` might trigger the default behavior of checking the environment.)